Don't call `setTheme` if `_theme` hasn't changed
`setTheme` can be an expensive operation because it parses the raw theme
every time it's called:
https://github.com/microsoft/vscode-textmate/blob/9e3c5941668cbfcee5095eaec0e58090fda8f316/src/main.ts#L93-L95

In https://github.com/banga/git-split-diffs/, I invoke
`codeToThemedTokens` many times, one line at a time, so the cost of
`setTheme` adds up. In my local profiling, this was taking about 25% of
the time spent in `codeToThemedTokens`. I realize this is probably not
the intended way to use this function, but there isn't an easy
alternative for syntax highlighting individual lines in a diff.